refactor: rename --> AppInsights reporter
