+++ packages/salesforcedx-utils-vscode/src/telemetry/appInsightsReporter.ts
